backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
pillow==10.1.0
pymupdf==1.23.8
google-generativeai==0.3.1
google-genai
python-dotenv==1.0.0
jinja2==3.1.2
pydantic==2.5.3
pydantic-settings==2.1.0
pymongo==4.6.1
motor==3.3.2
